refactor: dml과 구분 없이 사용하기 위해 패키지 구조 수정
