+++ src/main/java/persistence/entity/Person.java
@@ -18,4 +18,10 @@ public class Person {
     @Column(nullable = false)
     private String email;
 
+    public Person(Long id, String name, Integer age, String email) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.email = email;
+    }
 }
