--- src/main/java/persistence/sql/ddl/dialect/Dialect.java
@@ -1,11 +1,0 @@
-package persistence.sql.ddl.dialect;
-
-import persistence.sql.ddl.dialect.database.ConstraintsMapper;
-import persistence.sql.ddl.dialect.database.TypeMapper;
-
-public interface Dialect {
-
-    TypeMapper getTypeMapper();
-
-    ConstraintsMapper getConstantTypeMapper();
-}
