--- src/main/java/persistence/sql/ddl/dialect/database/ConstraintsMapper.java
@@ -1,9 +1,0 @@
-package persistence.sql.ddl.dialect.database;
-
-import persistence.sql.ddl.query.model.Constraints;
-
-public interface ConstraintsMapper {
-
-    String getConstantType(Constraints constantType);
-
-}
