+++ src/main/java/persistence/sql/dialect/Dialect.java
@@ -1,0 +1,11 @@
+package persistence.sql.dialect;
+
+import persistence.sql.dialect.database.ConstraintsMapper;
+import persistence.sql.dialect.database.TypeMapper;
+
+public interface Dialect {
+
+    TypeMapper getTypeMapper();
+
+    ConstraintsMapper getConstantTypeMapper();
+}
