+++ src/main/java/persistence/sql/dialect/database/ConstraintsMapper.java
@@ -1,0 +1,9 @@
+package persistence.sql.dialect.database;
+
+import persistence.sql.entity.model.Constraints;
+
+public interface ConstraintsMapper {
+
+    String getConstantType(Constraints constantType);
+
+}
