+++ src/main/java/persistence/sql/dialect/exception/InvalidJavaClassException.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.dialect.exception;
+package persistence.sql.dialect.exception;
 
 public class InvalidJavaClassException extends RuntimeException {
     private static final String MESSAGE = "지원하지 않는 타입이 들어왔습니다.";
