+++ src/main/java/persistence/sql/dialect/h2/H2Dialect.java
@@ -1,8 +1,8 @@
-package persistence.sql.ddl.dialect.h2;
+package persistence.sql.dialect.h2;
 
-import persistence.sql.ddl.dialect.Dialect;
-import persistence.sql.ddl.dialect.database.ConstraintsMapper;
-import persistence.sql.ddl.dialect.database.TypeMapper;
+import persistence.sql.dialect.Dialect;
+import persistence.sql.dialect.database.ConstraintsMapper;
+import persistence.sql.dialect.database.TypeMapper;
 
 public class H2Dialect implements Dialect {
     private final TypeMapper typeMapper;
