+++ src/main/java/persistence/sql/entity/model/DomainTypes.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.query.model;
+package persistence.sql.entity.model;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
