+++ src/test/java/persistence/sql/ddl/dialect/database/DatabaseTest.java
@@ -5,9 +5,10 @@ import database.H2;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.ddl.dialect.DialectResolution;
-import persistence.sql.ddl.dialect.DialectResolutionInfo;
-import persistence.sql.ddl.dialect.exception.NotFoundDatabase;
+import persistence.sql.dialect.DialectResolution;
+import persistence.sql.dialect.DialectResolutionInfo;
+import persistence.sql.dialect.database.Database;
+import persistence.sql.dialect.exception.NotFoundDatabase;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
