+++ src/test/java/persistence/sql/ddl/dialect/database/TypeMapperTest.java
@@ -6,10 +6,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import persistence.sql.ddl.dialect.Dialect;
-import persistence.sql.ddl.dialect.exception.InvalidJavaClassException;
-import persistence.sql.ddl.dialect.h2.H2Dialect;
-import persistence.sql.ddl.query.model.ColumnType;
+import persistence.sql.dialect.Dialect;
+import persistence.sql.dialect.database.TypeMapper;
+import persistence.sql.dialect.exception.InvalidJavaClassException;
+import persistence.sql.dialect.h2.H2Dialect;
+import persistence.sql.entity.model.ColumnType;
 
 import java.util.Optional;
 import java.util.stream.Stream;
